requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic_settings
aiogram
python-docx
aiosqlite
asyncpg
requests
urllib3
cryptography
aiohttp
PyJWT
gunicorn
alembic
apscheduler
python-dotenv
psycopg2-binary
passlib[bcrypt]
pytz
python-multipart
pytest
pytest-asyncio
